refactor: split triage observation helpers
